requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
ansible==9.5.1
ansible-core==2.16.5
ansible-pylibssh==1.1.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
Brotli==1.1.0
brotlicffi==1.1.0.0
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
configparser==6.0.1
cryptography==42.0.5
deepdiff==6.7.1
elementpath==4.4.0
et_xmlfile==1.1.0
frozenlist==1.4.1
future==1.0.0
geoip2==4.8.0
idna==3.7
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
junit-xml==1.9
jxmlease==1.0.3
lxml==5.1.0
MarkupSafe==2.1.5
maxminddb==2.6.1
multidict==6.0.5
ncclient==0.6.15
netaddr==1.2.1
netmiko==4.3.0
ntc_templates==4.4.0
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.1
packaging==24.0
paramiko==3.4.0
passlib==1.7.4
pexpect==4.9.0
poetry-core==1.9.0
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.0
pycares==4.4.0
pycparser==2.22
pycrypto==3.20.0
pycryptodome==3.20.0
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
resolvelib==1.0.1
rpds-py==0.17.1
scp==0.14.5
six==1.16.0
tabulate==0.9.0
textfsm==1.1.3
ttp==0.9.5
typing_extensions==4.11.0
urllib3==2.2.1
xmltodict==0.13.0
yangson==1.5.2
yarl==1.9.4
